chore: move tests from batch.rs into tests/mod.rs
